Avoid creating circular symlinks in launch scripts (#34)
The launcher script for csharp_binary and csharp_nunit_test rules
creates symlinks (ln -s -f) at the top level of runfiles for its
.exe and other files which normally live in subdirectories.

However, when the .exe was already at the top level, it would be
overwritten with a circular symlink. To fix this we check whether
the symlink is needed before creating it.