Use a "response file" to pass arguments to csc.
Windows has a max command-line length of around 32k bytes. The default
for Args is to spill to param files if the length of the executable,
params and spaces between them sum to that number. Unfortunately the
math doesn't work out exactly like that on Windows (e.g. there is also
a null terminator.) For now, setting use_always to True is the
conservative option. Long command lines are probable with C# due to
organizing files by namespace.