Adding support for correctly building .NET Standard assemblies
This change introduces a new rule type, netstandard_library, and
all of the supporting data structures required to build it (context,
SDK definition, toolchain, etc.). Using the netstandard_library rule,
it should now be possible to correctly build a .NET Standard 2.0
assembly, which can be used, in turn, as a reference in a .NET Core
project.

This commit only supports creating .NET Standard 2.0 or greater
assemblies.