Add Gitpod support
## What?
This adds Gitpod config for the repository.

## Why?
Gitpod allows developers to open the repository in ephemeral workspaces to explore the codebase or make changes.
This makes it easier for both curious developers and contributors to work with the repository without having to set up various tooling to get things working.